+++ src/atoms/x-alert/x-alert.vue
@@ -31,7 +31,7 @@ const handleClose = () => {
 <template>
   <v-alert
     v-bind="attrs"
-    :type="props.type"
+    :type="(props.type as any)"
     :title="props.title"
     :text="props.text"
     :closable="props.closable"
@@ -40,19 +40,19 @@ const handleClose = () => {
     @click:close="handleClose"
   >
     <template v-if="$slots.prepend" #prepend="scope">
-      <slot name="prepend" v-bind="scope || {}" />
+      <slot name="prepend" v-bind="(scope as any) || {}" />
     </template>
     <template v-if="$slots.title" #title="scope">
-      <slot name="title" v-bind="scope || {}" />
+      <slot name="title" v-bind="(scope as any) || {}" />
     </template>
     <template #default="scope">
-      <slot v-bind="scope || {}" />
+      <slot v-bind="(scope as any) || {}" />
     </template>
     <template v-if="$slots.append" #append="scope">
-      <slot name="append" v-bind="scope || {}" />
+      <slot name="append" v-bind="(scope as any) || {}" />
     </template>
     <template v-if="$slots.close" #close="scope">
-      <slot name="close" v-bind="scope || {}" />
+      <slot name="close" v-bind="(scope as any) || {}" />
     </template>
   </v-alert>
 </template>