+++ src/atoms/x-btn/x-btn.vue
@@ -49,16 +49,16 @@ const resolvedColor = computed(() => {
     :class="['x-btn', { 'x-btn--glass': isGlassVariant }]"
   >
     <template v-if="$slots.prepend" #prepend="scope">
-      <slot name="prepend" v-bind="scope || {}" />
+      <slot name="prepend" v-bind="(scope as any) || {}" />
     </template>
     <template v-if="$slots.append" #append="scope">
-      <slot name="append" v-bind="scope || {}" />
+      <slot name="append" v-bind="(scope as any) || {}" />
     </template>
     <template v-if="$slots.loader" #loader="scope">
-      <slot name="loader" v-bind="scope || {}" />
+      <slot name="loader" v-bind="(scope as any) || {}" />
     </template>
     <template #default="scope">
-      <slot v-bind="scope || {}" />
+      <slot v-bind="(scope as any) || {}" />
     </template>
   </v-btn>
 </template>