+++ src/atoms/x-dialog/types.d.ts
@@ -5,8 +5,10 @@ export interface XDialogProps {
   persistent?: boolean;
   scrollable?: boolean;
   fullscreen?: boolean;
+  transition?: string;
 }
 
 export interface XDialogEmits {
   (e: 'update:modelValue', value: boolean): void;
+  (e: 'update:model-value', value: boolean): void;
 }