+++ src/atoms/x-dialog/x-dialog.vue
@@ -21,7 +21,10 @@ const attrs = useAttrs();
 
 const isOpen = computed({
   get: () => props.modelValue,
-  set: (val: boolean) => emit('update:modelValue', val),
+  set: (val: boolean) => {
+    emit('update:modelValue', val);
+    emit('update:model-value', val);
+  },
 });
 </script>
 
@@ -34,24 +37,25 @@ const isOpen = computed({
     :persistent="props.persistent"
     :scrollable="props.scrollable"
     :fullscreen="props.fullscreen"
+    :transition="props.transition"
     class="x-dialog"
   >
     <template v-if="$slots.activator" #activator="scope">
-      <slot name="activator" v-bind="scope || {}" />
+      <slot name="activator" v-bind="(scope as any) || {}" />
     </template>
 
     <v-card class="x-dialog__surface">
       <template v-if="$slots.title" #title="scope">
-        <slot name="title" v-bind="scope || {}" />
+        <slot name="title" v-bind="(scope as any) || {}" />
       </template>
 
       <template #default="scope">
-        <slot v-bind="scope || {}" />
+        <slot v-bind="(scope as any) || {}" />
       </template>
 
       <template v-if="$slots.actions" #actions="scope">
         <div class="x-dialog__actions">
-          <slot name="actions" v-bind="scope || {}" />
+          <slot name="actions" v-bind="(scope as any) || {}" />
         </div>
       </template>
     </v-card>