+++ src/atoms/x-list-item/index.ts
@@ -1,0 +1,3 @@
+export { default as XListItem } from './x-list-item.vue';
+export * from './types.d';
+export * from './x-list-item.controller';