+++ src/atoms/x-list/index.ts
@@ -1,0 +1,3 @@
+export { default as XList } from './x-list.vue';
+export * from './types.d';
+export * from './x-list.controller';